Separate QgsFileWidget from QgsExternalStorageFileWidget
