+++ python/gui/auto_generated/qgsfilewidget.sip.in
@@ -79,104 +79,6 @@ Sets the current file ``path``.
 Sets whether the widget should be read only.
 %End
 
-    void setStorageType( const QString &storageType );
-%Docstring
-Set ``storageType`` storage type unique identifier as defined in :py:class:`QgsExternalStorageRegistry` or
-null QString if there is no storage defined.
-If no external storage has been defined, QgsFileWidget will only update file path according to
-selected files.
-
-.. seealso:: :py:func:`storageType`
-
-.. versionadded:: 3.22
-%End
-
-    QString storageType() const;
-%Docstring
-Returns storage type unique identifier as defined in :py:class:`QgsExternalStorageRegistry`.
-Returns null QString if there is no storage defined, only file selection.
-
-.. seealso:: :py:func:`setStorageType`
-
-.. versionadded:: 3.22
-%End
-
-    QgsExternalStorage *externalStorage() const;
-%Docstring
-Returns external storage used to store selected file names, None if none have been defined.
-If no external storage has been defined, QgsFileWidget will only update file path according to
-selected files.
-
-.. seealso:: :py:func:`setStorageType`
-
-.. versionadded:: 3.22
-%End
-
-    void setStorageAuthConfigId( const QString &authCfg );
-%Docstring
-Sets the authentication configuration ID to be used for the current external storage (if
-defined)
-
-.. versionadded:: 3.22
-%End
-
-    const QString &storageAuthConfigId() const;
-%Docstring
-Returns the authentication configuration ID used for the current external storage (if defined)
-
-.. versionadded:: 3.22
-%End
-
-    void setStorageUrlExpression( const QString &urlExpression );
-%Docstring
-Set ``urlExpression`` expression, which once evaluated, provide the URL used to store selected
-documents. This is used only if an external storage has been defined
-
-.. seealso:: :py:func:`setStorageType`
-
-.. versionadded:: 3.22
-%End
-
-    QString storageUrlExpressionString() const;
-%Docstring
-Returns the original, unmodified expression string, which once evaluated, provide the
-URL used to store selected documents. This is used only if an external storage has been defined.
-Returns null if no expression has been set.
-
-.. seealso:: :py:func:`setStorageUrlExpression`
-
-.. versionadded:: 3.22
-%End
-
-    QgsExpression *storageUrlExpression() const;
-%Docstring
-Returns expression, which once evaluated, provide the URL used to store selected
-documents. This is used only if an external storage has been defined.
-Returns null if no expression has been set.
-
-.. seealso:: :py:func:`setStorageUrlExpression`
-
-.. versionadded:: 3.22
-%End
-
-    void setExpressionContext( const QgsExpressionContext &context );
-%Docstring
-Set expression context to be used when for storage URL expression evaluation
-
-.. seealso:: :py:func:`setStorageUrlExpression`
-
-.. versionadded:: 3.22
-%End
-
-    const QgsExpressionContext &expressionContext() const;
-%Docstring
-Returns expression context used for storage url expression evaluation
-
-.. seealso:: :py:func:`storageUrlExpression`
-
-.. versionadded:: 3.22
-%End
-
     QString dialogTitle() const;
 %Docstring
 Returns the open file dialog title.
@@ -352,31 +254,33 @@ the appearance and behavior of the line edit portion of the widget.
 .. versionadded:: 3.0
 %End
 
-    void setMessageBar( QgsMessageBar *messageBar );
-%Docstring
-Set ``messageBar`` to report messages
-
-.. versionadded:: 3.22
-%End
-
-    QgsMessageBar *messageBar() const;
-%Docstring
-Returns message bar used to report messages
-
-.. versionadded:: 3.22
-%End
-
-    static QgsExpressionContextScope *createFileWidgetScope();
-%Docstring
-Creates and Returns an expression context scope specific to QgsFileWidget
-It defines the variable containing the user selected file name
-%End
-
   signals:
 
     void fileChanged( const QString &path );
 %Docstring
 Emitted whenever the current file or directory ``path`` is changed.
+%End
+
+  protected:
+
+    virtual void updateLayout();
+
+    virtual void setSelectedFileNames( QStringList fileNames );
+
+    static bool isMultiFiles( const QString &path );
+
+    void setFilePaths( const QStringList &filePaths );
+
+
+
+    QString toUrl( const QString &path ) const;
+%Docstring
+returns a HTML code with a link to the given file path
+%End
+
+    QString relativePath( const QString &filePath, bool removeRelative ) const;
+%Docstring
+Returns a filePath with relative path options applied (or not) !
 %End
 
 };
